move system_cmd to its own cmd dir
